# api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: SGDyPA Backend API
  version: 0.1.0
  summary: Contract-first API surface for the SGDyPA SPA and trusted integrations.
  description: |
    Versioned OpenAPI contract for `/api/v1/`. New endpoints must update this
    file in the same pull request as their implementation.
...
servers:
  - url: /api/v1
    description: Versioned API root
security:
  - bearerJwt: []
...
paths:
...
  /roles:
    get:
      tags: [Identity]
      operationId: listRoles
      summary: List role definitions.
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Roles.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Identity]
      operationId: createRole
      summary: Create an org-scoped role definition.
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleWrite'
      responses:
        '201':
          description: Role created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued access token validated by the API.
  parameters:
    OrganizationIdHeader:
      name: X-Organization-Id
      in: header
      required: true
      description: Active organization scope validated against the caller memberships.
      schema:
        type: string
        format: uuid
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable UUID used to make mutating requests retry-safe.
      schema:
        type: string
        format: uuid
  responses:
    ErrorResponse:
      description: Stable API error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    Role:
      type: object
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        capabilities:
          type: array
          items: {type: string}
        is_system: {type: boolean}
    RoleWrite:
      type: object
      required: [code, name]
      properties:
        code: {type: string}
        name: {type: string}
...
    ErrorCode:
      type: string
      enum:
        - scope_frozen
        - illegal_transition
        - legal_hold_active
        - self_approval_forbidden
        - validation_failed
        - stale_state
        - idempotency_key_required
        - idempotency_key_invalid
        - idempotency_key_conflict
        - authentication_failed
        - not_authenticated
        - not_found
        - method_not_allowed
        - permission_denied
        - parse_error
        - throttled
        - unsupported_media_type
        - internal_error
    ErrorDetail:
      type: object
      additionalProperties: true
      properties:
        field:
          type: string
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
    ErrorDetails:
      description: >-
        Error context. An array of ErrorDetail entries for most errors, or a
        map of field name to validation messages for validation_failed
        responses.
      oneOf:
        - type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        - type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details, request_id]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            request_id:
              type: string
              description: Correlation id echoed from the X-Request-Id middleware.
            details:
              $ref: '#/components/schemas/ErrorDetails'